refactor: consolidate BattlEye logic
